moved torch wrapper to library wrapper folder along with cpp test
